requirements.txt
flask==2.0.1
numpy==1.22.4
requests==2.28.1
pyproj==3.3.1
python-dateutil==2.8.2 